JQuery 'ready' should be fired after devextreme 'ready'
