+++ js/core/utils/dom.js
@@ -169,20 +169,6 @@ var contains = function(container, element) {
     return container.nodeType === Node.DOCUMENT_NODE ? container.body.contains(element) : container.contains(element);
 };
 
-exports.ready = function(callback) {
-    //NOTE: we can't use document.readyState === "interactive" because of ie9/ie10 support
-    if(document.readyState === "complete" || (document.readyState !== "loading" && !document.documentElement.doScroll)) {
-        callback();
-        return;
-    }
-
-    var loadedCallback = function() {
-        callback();
-        document.removeEventListener("DOMContentLoaded", loadedCallback);
-    };
-    document.addEventListener("DOMContentLoaded", loadedCallback);
-};
-
 exports.resetActiveElement = resetActiveElement;
 exports.createMarkupFromString = createMarkupFromString;
 exports.triggerShownEvent = triggerVisibilityChangeEvent("dxshown");
