+++ js/integration/jquery.js
@@ -11,7 +11,7 @@ if(jQuery) {
     }
 
     require("./jquery/renderer");
-    require("./jquery/themes");
+    require("./jquery/hold_ready");
     require("./jquery/events");
     require("./jquery/easing");
     require("./jquery/element_data");
