+++ js/ui/themes.js
@@ -3,6 +3,7 @@
 var $ = require("../core/renderer"),
     errors = require("./widget/ui.errors"),
     domUtils = require("../core/utils/dom"),
+    ready = require("../core/utils/ready"),
     each = require("../core/utils/iterator").each,
     devices = require("../core/devices"),
     viewPortUtils = require("../core/utils/view_port"),
@@ -296,7 +297,7 @@ init({
     }
 });
 
-domUtils.ready(function() {
+ready(function() {
     if($(DX_LINK_SELECTOR, context).length) {
         throw errors.Error("E0022");
     }
